fix: keep resize input uint8 to match the HF processor's antialias kernel
F.interpolate's antialias bicubic has two kernels: uint8 input selects
the PIL-equivalent kernel that the HF processor's resize uses, while
float input selects a different one. Converting to float before the
resize therefore picked the wrong kernel and diverged from the stock
path by up to ~24 LSB on ~23% of pixels for non-aligned sizes.

Keep the tensor uint8 through the interpolate and convert to float
afterwards; clamp/round are retained as a no-op safety net for
backends that return float from uint8 input.

Verified on CPU: the resize is now bit-exact with the HF processor
across 12 shapes (identity/upscale/downscale), and the end-to-end
pixel_values (through patchify and the FusedInputNorm affine) differ
from the HF pipeline by at most 2.4e-7 (fp32 last-ulp rounding from
multiply-add association, far below bf16 resolution).

+++ vllm_ascend/models/qwen3_5_vl.py
@@ -136,10 +136,12 @@ class _AscendVLPreprocessMixin:
         chunks = flat.split([c * h * w for h, w in hw_list])
         patches = []
         for chunk, (h, w), (_, gh, gw) in zip(chunks, hw_list, grid_thw.tolist()):
-            img = chunk.reshape(1, c, h, w).float()
-            # bicubic + antialias reproduces HF/PIL resize (cosine sim 0.999996)
+            # Keep uint8 through the resize: this selects the antialias kernel
+            # the HF processor uses (float input takes a different kernel);
+            # clamp/round stay as a no-op on the already-quantized output.
+            img = chunk.reshape(1, c, h, w)
             img = F.interpolate(img, size=[gh * ps, gw * ps], mode="bicubic", align_corners=False, antialias=True)
-            img = img.clamp(0, 255).round().squeeze(0)
+            img = img.clamp(0, 255).round().float().squeeze(0)
             # patchify: (C, rh, rw) -> (gh*gw, C*tps*ps*ps), HF layout.
             # Drop the batch dim before unsqueeze/expand: keeping it (as HF's
             # batched patchify does) creates a 9-D tensor, which CANN rejects
